requirements.txt
azure-cognitiveservices-speech==1.34.0
openai==1.3.0
google-cloud-speech==2.21.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
python-dotenv==1.0.0
pydub==0.25.1
whisper==1.1.10
torch==2.0.1
torchaudio==2.0.2
transformers==4.30.2
datasets==2.14.0
evaluate==0.4.0
jiwer==3.0.3 